requirements.txt
# Add your Python dependencies here
google-generativeai
python-dotenv
openai